backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.30
alembic==1.13.1
redis[hiredis]==5.0.4
nats-py==2.7.2
pydantic==2.7.1
pydantic-settings==2.2.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
docker==7.0.0
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-exporter-otlp==1.24.0
prometheus-fastapi-instrumentator==7.0.0
structlog==24.2.0
psutil==5.9.8
aiofiles==23.2.1
cryptography==42.0.8
authlib==1.3.0
email-validator==2.1.1
pgvector==0.3.2
stripe==9.9.0
setuptools